requirements.txt
gspread==5.10.0
oauth2client==4.1.3
pandas==1.5.3
matplotlib==3.7.1
python-dotenv==1.0.0
gspread-formatting==1.1.2
numpy==1.24.3